+++ src/lib/Scenes/City/City.tsx
@@ -1,33 +1,60 @@
 import { Box, color, Flex, Theme } from "@artsy/palette"
+import { Schema, screenTrack } from "lib/utils/track"
 import React, { Component } from "react"
 import { NativeModules, ScrollView } from "react-native"
 import { RelayProp } from "react-relay"
 import styled from "styled-components/native"
-import { BucketKey, BucketResults } from "../Map/Bucket"
+import { BucketKey, BucketResults } from "../Map/bucketCityResults"
 import { FiltersBar } from "../Map/Components/FiltersBar"
 import { EventEmitter } from "../Map/EventEmitter"
-import { MapTab as Tab } from "../Map/types"
+import { MapTab } from "../Map/types"
+import { cityTabs } from "./cityTabs"
 import { AllEvents } from "./Components/AllEvents"
 import { EventList } from "./Components/EventList"
 
 interface Props {
   verticalMargin?: number
   isDrawerOpen?: boolean
+  citySlug: string
+  tracking: any
 }
 
 interface State {
   buckets?: BucketResults
-  filter: Tab
+  filter: MapTab
   relay: RelayProp
   cityName: string
   citySlug: string
   sponsoredContent: { introText: string; artGuideUrl: string }
 }
 
+const screenSchemaForCurrentTabState = currentSelectedTab => {
+  switch (currentSelectedTab) {
+    case "all":
+      return Schema.PageNames.CityGuideAllGuide
+    case "saved":
+      return Schema.PageNames.CityGuideSavedGuide
+    case "fairs":
+      return Schema.PageNames.CityGuideFairsGuide
+    case "galleries":
+      return Schema.PageNames.CityGuideGalleriesGuide
+    case "museums":
+      return Schema.PageNames.CityGuideMuseumsGuide
+    default:
+      return null
+  }
+}
+
+@screenTrack<Props>(props => ({
+  context_screen: screenSchemaForCurrentTabState("all"),
+  context_screen_owner_type: Schema.OwnerEntityTypes.CityGuide,
+  context_screen_owner_slug: props.citySlug,
+  context_screen_owner_id: props.citySlug,
+}))
 export class CityView extends Component<Props, State> {
   state = {
     buckets: null,
-    filter: { id: "all", text: "All events" },
+    filter: cityTabs[0],
     relay: null,
     cityName: "",
     citySlug: "",
@@ -35,14 +62,6 @@ export class CityView extends Component<Props, State> {
   }
   scrollViewVerticalStart = 0
   scrollView: ScrollView = null
-
-  filters: Tab[] = [
-    { id: "all", text: "All" },
-    { id: "saved", text: "Saved" },
-    { id: "fairs", text: "Fairs" },
-    { id: "galleries", text: "Galleries" },
-    { id: "museums", text: "Museums" },
-  ]
 
   handleEvent = ({
     filter,
@@ -52,7 +71,7 @@ export class CityView extends Component<Props, State> {
     relay,
     sponsoredContent,
   }: {
-    filter: Tab
+    filter: MapTab
     buckets: BucketResults
     cityName: string
     relay: RelayProp
@@ -77,6 +96,12 @@ export class CityView extends Component<Props, State> {
     EventEmitter.unsubscribe("map:change", this.handleEvent)
   }
 
+  componentWillReceiveProps(nextProps) {
+    if (this.props.isDrawerOpen !== nextProps.isDrawerOpen) {
+      this.fireScreenViewAnalytics()
+    }
+  }
+
   componentDidUpdate() {
     if (!this.props.isDrawerOpen && this.scrollView) {
       this.scrollView.scrollTo({ x: 0, y: 0, animated: true })
@@ -86,6 +111,15 @@ export class CityView extends Component<Props, State> {
       // We have the Relay response; post a notification so that the ARMapContainerViewController can finalize the native UI.
       NativeModules.ARNotificationsManager.postNotificationName("ARLocalDiscoveryQueryResponseReceived", {})
     }
+  }
+
+  fireScreenViewAnalytics = () => {
+    this.props.tracking.trackEvent({
+      context_screen: screenSchemaForCurrentTabState(this.state.filter.id),
+      context_screen_owner_type: Schema.OwnerEntityTypes.CityGuide,
+      context_screen_owner_slug: this.state.citySlug,
+      context_screen_owner_id: this.state.citySlug,
+    })
   }
 
   render() {
@@ -102,7 +136,7 @@ export class CityView extends Component<Props, State> {
                 <Handle />
               </Flex>
               <FiltersBar
-                tabs={this.filters}
+                tabs={cityTabs}
                 goToPage={activeIndex => EventEmitter.dispatch("filters:change", activeIndex)}
               />
               <ScrollView