+++ src/lib/Scenes/City/Components/BMWEventSection/__tests__/index-tests.tsx
@@ -6,18 +6,16 @@ import { BMWEventSection } from "../index"
 
 const data = [
   {
-    node: {
-      name: "PALAY, Trapunto Murals by Pacita Abad",
-      __id: "U2hvdzpwYWNpdGEtYWJhZC1hcnQtZXN0YXRlLXBhbGF5LXRyYXB1bnRvLW11cmFscy1ieS1wYWNpdGEtYWJhZA==",
-      id: "pacita-abad-art-estate-palay-trapunto-murals-by-pacita-abad",
-      cover_image: {
-        url: "",
-      },
-      end_at: "2001-12-15T12:00:00+00:00",
-      start_at: "2001-11-12T12:00:00+00:00",
-      partner: {
-        name: "Pacita Abad Art Estate",
-      },
+    name: "PALAY, Trapunto Murals by Pacita Abad",
+    __id: "U2hvdzpwYWNpdGEtYWJhZC1hcnQtZXN0YXRlLXBhbGF5LXRyYXB1bnRvLW11cmFscy1ieS1wYWNpdGEtYWJhZA==",
+    id: "pacita-abad-art-estate-palay-trapunto-murals-by-pacita-abad",
+    cover_image: {
+      url: "",
+    },
+    end_at: "2001-12-15T12:00:00+00:00",
+    start_at: "2001-11-12T12:00:00+00:00",
+    partner: {
+      name: "Pacita Abad Art Estate",
     },
   },
 ]
@@ -34,6 +32,7 @@ describe("CityEvent", () => {
           }}
           relay={{ environment: {} } as RelayProp}
           data={data}
+          citySlug={"new-york-us"}
         />
       </Theme>
     )