+++ src/lib/Scenes/Map/MapDataSource.tsx
@@ -1,0 +1,7 @@
+import React, { Component } from "react"
+
+export class MapDataSource extends Component<any> {
+  render() {
+    return null
+  }
+}