+++ src/lib/relay/QueryRenderers.tsx
@@ -30,12 +30,13 @@ import Inbox from "../Containers/Inbox"
 Inbox
 // tslint:enable:no-unused-expression
 
+import { EventStatus } from "__generated__/CitySectionListQuery.graphql"
 import { QueryRenderersArtistQuery } from "__generated__/QueryRenderersArtistQuery.graphql"
 import { QueryRenderersBidFlowQuery } from "__generated__/QueryRenderersBidFlowQuery.graphql"
 import { QueryRenderersCityBMWListQuery } from "__generated__/QueryRenderersCityBMWListQuery.graphql"
 import { QueryRenderersCityFairListQuery } from "__generated__/QueryRenderersCityFairListQuery.graphql"
-import { PartnerShowPartnerType } from "__generated__/QueryRenderersCitySectionListQuery.graphql"
 import { QueryRenderersCitySectionListQuery } from "__generated__/QueryRenderersCitySectionListQuery.graphql"
+import { PartnerShowPartnerType } from "__generated__/QueryRenderersCitySectionListQuery.graphql"
 import { QueryRenderersConversationQuery } from "__generated__/QueryRenderersConversationQuery.graphql"
 import { QueryRenderersFairQuery } from "__generated__/QueryRenderersFairQuery.graphql"
 import { QueryRenderersForYouQuery } from "__generated__/QueryRenderersForYouQuery.graphql"
@@ -50,7 +51,7 @@ import { QueryRenderersShowArtworksQuery } from "__generated__/QueryRenderersSho
 import { QueryRenderersShowMoreInfoQuery } from "__generated__/QueryRenderersShowMoreInfoQuery.graphql"
 import { QueryRenderersShowQuery } from "__generated__/QueryRenderersShowQuery.graphql"
 import { QueryRenderersWorksForYouQuery } from "__generated__/QueryRenderersWorksForYouQuery.graphql"
-import { BucketKey } from "lib/Scenes/Map/Bucket"
+import { BucketKey } from "lib/Scenes/Map/bucketCityResults"
 import createEnvironment from "./createEnvironment"
 const environment = createEnvironment()
 
@@ -486,7 +487,12 @@ interface CitySectionListProps extends RendererProps {
   section: BucketKey
 }
 export const CitySectionListRenderer: React.SFC<CitySectionListProps> = ({ render, citySlug, section }) => {
-  const variables: { citySlug: string; partnerType?: PartnerShowPartnerType } = { citySlug }
+  const variables: {
+    citySlug: string
+    partnerType?: PartnerShowPartnerType
+    status?: EventStatus
+    dayThreshold?: number
+  } = { citySlug }
 
   switch (section) {
     case "museums":
@@ -495,15 +501,27 @@ export const CitySectionListRenderer: React.SFC<CitySectionListProps> = ({ rende
     case "galleries":
       variables.partnerType = "GALLERY"
       break
+    case "closing":
+      variables.status = "CLOSING_SOON"
+      variables.dayThreshold = 7
+      break
+    case "opening":
+      variables.status = "UPCOMING"
+      variables.dayThreshold = 7
   }
 
   return (
     <QueryRenderer<QueryRenderersCitySectionListQuery>
       environment={environment}
       query={graphql`
-        query QueryRenderersCitySectionListQuery($citySlug: String!, $partnerType: PartnerShowPartnerType) {
+        query QueryRenderersCitySectionListQuery(
+          $citySlug: String!
+          $partnerType: PartnerShowPartnerType
+          $status: EventStatus
+          $dayThreshold: Int
+        ) {
           city(slug: $citySlug) {
-            ...CitySectionList_city @arguments(partnerType: $partnerType)
+            ...CitySectionList_city @arguments(partnerType: $partnerType, status: $status, dayThreshold: $dayThreshold)
           }
         }
       `}