+++ src/lib/utils/convertCityToGeoJSON.ts
@@ -1,4 +1,6 @@
-import { MapGeoFeatureCollection } from "lib/Scenes/Map/types"
+import { GlobalMap_viewer } from "__generated__/GlobalMap_viewer.graphql"
+
+import { Fair, MapGeoFeatureCollection, Show } from "lib/Scenes/Map/types"
 
 // Here is a sample GeoJSON document
 // {
@@ -16,11 +18,29 @@ import { MapGeoFeatureCollection } from "lib/Scenes/Map/types"
 //       },
 // }
 
+export type FairsEdge = GlobalMap_viewer["city"]["fairs"]["edges"]
+export type ShowsEdge = GlobalMap_viewer["city"]["shows"]["edges"]
+
+export const showsToGeoCityShow = (edges: Show[]): Show[] =>
+  edges.map(node => {
+    return {
+      ...node,
+      icon: node.is_followed ? "pin-saved" : "pin",
+    }
+  })
+
+export const fairToGeoCityFairs = (edges: Fair[]): Fair[] =>
+  edges.map(node => {
+    return {
+      ...node,
+      icon: "pin",
+    }
+  })
+
 export const convertCityToGeoJSON = data => {
   return {
     type: "FeatureCollection",
     features: data
-      .map(({ node }) => node)
       // The API has (at least once) given us back shows without locations
       // so we should protect against runtime errors.
       .filter(feature => feature.location && feature.location.coordinates)