+++ src/lib/utils/track.ts
@@ -88,6 +88,19 @@ export namespace Schema {
     BidFlowRegistrationResultConfirmed = "RegistrationConfirmed",
     BidFlowRegistrationResultPending = "RegistrationPending",
     BidFlowRegistrationResultError = "RegistrationError",
+    CityGuideAllGuide = "CityGuideAllGuide",
+    CityGuideAllMap = "CityGuideAllMap",
+    CityGuideSavedGuide = "CityGuideSavedGuide",
+    CityGuideSavedMap = "CityGuideSavedMap",
+    CityGuideFairsGuide = "CityGuideFairsGuide",
+    CityGuideFairsMap = "CityGuideFairsMap",
+    CityGuideGalleriesGuide = "CityGuideGalleriesGuide",
+    CityGuideGalleriesMap = "CityGuideGalleriesMap",
+    CityGuideMuseumsGuide = "CityGuideMuseumsGuide",
+    CityGuideMuseumsMap = "CityGuideMuseumsMap",
+    CityPicker = "CityPicker",
+    ClusteredMapPin = "ClusteredMapPin",
+    SingleMapPin = "SingleMapPin",
     ConversationPage = "Conversation",
     ConsignmentsWelcome = "ConsignmentsWelcome",
     ConsignmentsOverView = "ConsignmentsOverview",
@@ -100,7 +113,7 @@ export namespace Schema {
     FairAllArtistsPage = "FairAllArtistsPage",
     FairAllArtworksPage = "FairAllArtworksPage",
     FairAllExhibitorsPage = "FairAllExhibitorsPage",
-    FairBMWArtActivationPage = "FairBMWArtActivationPage",
+    BMWFairActivation = "BMWFairActivation",
     FairBoothPage = "FairBoothPage",
     InboxPage = "Inbox",
     InquiryPage = "Inquiry",
@@ -115,6 +128,7 @@ export namespace Schema {
   export enum OwnerEntityTypes {
     Artist = "Artist",
     Artwork = "Artwork",
+    CityGuide = "CityGuide",
     Conversation = "Conversation",
     Gallery = "Gallery",
     Gene = "Gene",
@@ -149,6 +163,12 @@ export namespace Schema {
     ArtistUnfollow = "artistUnfollow",
     ArtistWorks = "artistWorks",
     ArtistShows = "artistShows",
+
+    /**
+     * City and Map Page Events
+     */
+
+    GetBMWArtGuide = "getBMWArtGuide",
 
     /**
      * Gene Page Events
@@ -222,7 +242,7 @@ export namespace Schema {
     GalleryFollow = "galleryFollow",
     GalleryUnfollow = "galleryUnfollow",
     Search = "search",
-    ViewBMWPressRelease = "pressRelease",
+    PressRelease = "pressRelease",
 
     /**
      * Fair and show shared page events